Flytt tester av satstypeTidslinje til SatsServiceTest.kt
